+++ src/Behat/Behat/Context/ContextFactory.php
@@ -68,16 +68,18 @@ final class ContextFactory
     /**
      * Creates and initializes context class.
      *
-     * @param string $class
-     * @param array  $arguments
+     * @param string             $class
+     * @param array              $arguments
+     * @param ArgumentResolver[] $singleUseResolvers
      *
      * @return Context
      */
-    public function createContext($class, array $arguments = array())
+    public function createContext($class, array $arguments = array(), array $singleUseResolvers = array())
     {
         $reflection = new ReflectionClass($class);
-        $arguments = $this->createArguments($reflection, $arguments);
-        $context = $this->createInstance($reflection, $arguments);
+        $resolvers = array_merge($singleUseResolvers, $this->argumentResolvers);
+        $resolvedArguments = $this->resolveArguments($reflection, $arguments, $resolvers);
+        $context = $this->createInstance($reflection, $resolvedArguments);
         $this->initializeInstance($context);
 
         return $context;
@@ -86,14 +88,15 @@ final class ContextFactory
     /**
      * Resolves arguments for a specific class using registered argument resolvers.
      *
-     * @param ReflectionClass $reflection
-     * @param array           $arguments
+     * @param ReflectionClass    $reflection
+     * @param array              $arguments
+     * @param ArgumentResolver[] $resolvers
      *
      * @return mixed[]
      */
-    private function createArguments(ReflectionClass $reflection, array $arguments)
+    private function resolveArguments(ReflectionClass $reflection, array $arguments, array $resolvers)
     {
-        foreach ($this->argumentResolvers as $resolver) {
+        foreach ($resolvers as $resolver) {
             $arguments = $resolver->resolveArguments($reflection, $arguments);
         }
 
@@ -112,7 +115,7 @@ final class ContextFactory
      * @param ReflectionClass $reflection
      * @param array           $arguments
      *
-     * @return Context
+     * @return mixed
      */
     private function createInstance(ReflectionClass $reflection, array $arguments)
     {
@@ -127,8 +130,6 @@ final class ContextFactory
      * Initializes context class and returns new context instance.
      *
      * @param Context $context
-     *
-     * @return Context
      */
     private function initializeInstance(Context $context)
     {