+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [3.3.0] - 2016-12-25
+### Added
+  * [#973](https://github.com/Behat/Behat/pull/974): Added helper containers
+  * [#973](https://github.com/Behat/Behat/pull/974): Added `SuiteScopedResolverFactory` extension point
+
 ## [3.2.3] - 2016-12-25
 ### Fixed
   * [#971](https://github.com/Behat/Behat/pull/971): Added support for suite names with hyphens
@@ -788,7 +793,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
   * Initial release
 
-[Unreleased]: https://github.com/Behat/Behat/compare/v3.2.3...HEAD
+[Unreleased]: https://github.com/Behat/Behat/compare/v3.3.0...HEAD
+[3.3.0]: https://github.com/Behat/Behat/compare/v3.2.3...v3.3.0
 [3.2.3]: https://github.com/Behat/Behat/compare/v3.2.2...v3.2.3
 [3.2.2]: https://github.com/Behat/Behat/compare/v3.2.1...v3.2.2
 [3.2.1]: https://github.com/Behat/Behat/compare/v3.2.0...v3.2.1